Improve schema types
