+++ packages/graphql/src/auraApi/schema/AuraSchemaGenerator.ts
@@ -1,55 +1,29 @@
 import type { GraphQLSchema } from "graphql";
-import { type ObjectTypeComposer } from "graphql-compose";
-import { Memoize } from "typescript-memoize";
 import type { Neo4jGraphQLSchemaModel } from "../../schema-model/Neo4jGraphQLSchemaModel";
 import type { ConcreteEntity } from "../../schema-model/entity/ConcreteEntity";
+import { AuraEntityOperations } from "../AuraEntityOperations";
 import { generateReadResolver } from "../resolvers/readResolver";
-import { EntitySchemaTypes } from "./EntitySchemaTypes";
 import { SchemaBuilder } from "./SchemaBuilder";
-
-export type StaticTypes = Record<"pageInfo", ObjectTypeComposer>;
+import { StaticTypes } from "./schema-types/StaticTypes";
+import { TopLevelEntityTypes } from "./schema-types/TopLevelEntityTypes";
 
 export class AuraSchemaGenerator {
     private schemaBuilder: SchemaBuilder;
-    private entityTypes: Map<ConcreteEntity, EntitySchemaTypes>;
 
-    constructor(schemaModel: Neo4jGraphQLSchemaModel) {
+    constructor() {
         this.schemaBuilder = new SchemaBuilder();
-        this.entityTypes = this.generateEntityTypes(schemaModel);
     }
 
-    public generate(): GraphQLSchema {
-        this.createQueryFields();
+    public generate(schemaModel: Neo4jGraphQLSchemaModel): GraphQLSchema {
+        const staticTypes = new StaticTypes({ schemaBuilder: this.schemaBuilder });
+        const entityTypes = this.generateEntityTypes(schemaModel, staticTypes);
+        this.createQueryFields(entityTypes);
 
         return this.schemaBuilder.build();
     }
 
-    @Memoize()
-    private get staticTypes(): StaticTypes {
-        return {
-            pageInfo: this.createPageInfoType(),
-        } as const;
-    }
-
-    private generateEntityTypes(schemaModel: Neo4jGraphQLSchemaModel): Map<ConcreteEntity, EntitySchemaTypes> {
-        const resultMap = new Map<ConcreteEntity, EntitySchemaTypes>();
-        for (const entity of schemaModel.entities.values()) {
-            if (entity.isConcreteEntity()) {
-                const entitySchemaTypes = new EntitySchemaTypes({
-                    entity,
-                    schemaBuilder: this.schemaBuilder,
-                    staticTypes: this.staticTypes,
-                });
-
-                resultMap.set(entity, entitySchemaTypes);
-            }
-        }
-
-        return resultMap;
-    }
-
-    private createQueryFields(): void {
-        this.entityTypes.forEach((entitySchemaTypes, entity) => {
+    private createQueryFields(entityTypes: Map<ConcreteEntity, TopLevelEntityTypes>): void {
+        entityTypes.forEach((entitySchemaTypes, entity) => {
             const resolver = generateReadResolver({
                 entity,
             });
@@ -61,7 +35,24 @@ export class AuraSchemaGenerator {
         });
     }
 
-    private createPageInfoType(): ObjectTypeComposer {
-        return this.schemaBuilder.createObjectType("PageInfo", { hasNextPage: "Boolean", hasPreviousPage: "Boolean" });
+    private generateEntityTypes(
+        schemaModel: Neo4jGraphQLSchemaModel,
+        staticTypes: StaticTypes
+    ): Map<ConcreteEntity, TopLevelEntityTypes> {
+        const resultMap = new Map<ConcreteEntity, TopLevelEntityTypes>();
+        for (const entity of schemaModel.entities.values()) {
+            if (entity.isConcreteEntity()) {
+                const entitySchemaTypes = new TopLevelEntityTypes({
+                    entity,
+                    schemaBuilder: this.schemaBuilder,
+                    staticTypes,
+                    entityOperations: new AuraEntityOperations(entity),
+                });
+
+                resultMap.set(entity, entitySchemaTypes);
+            }
+        }
+
+        return resultMap;
     }
 }
